Process-oriented first pass.
This version uses a CSP-like structure, or process-oriented structure,
to handle concurrent execution/monitoring.

A 1-minute ticker process fires... ticks every minute.

Every tick, we monitor for 45 seconds.

Every 'n' ticks, the monitoring results are consolidated, and
then reported to the DB.

The DB is then cleared, and the counting starts again.