Data clearing properly
This fixes the local data storage. I need to make a list of all the things that
need to be considered here... because we're spawning UNIX processes for this work,
I have a lot of memory leak concerns. I'm *reasonably* confident in Racket's memory handling,
but I'm *not* confident about most things in Python.

That is, I'm more likely to be comfortable with a long-running Racket process than I am with a
long-running Python process. However, I don't have a lot of good reasons for that, save
for the fact that Racket... was basically designed as an OS.